refact(transition): rename modules
